Create | Check isAdmin about DELETE & PUT methods at serverside

+++ pages/api/products/[id].js
@@ -3,7 +3,9 @@ import Product from "../../../models/Product"
 
 export default async function handler(req, res) {
 
-    const { method, query: { id } } = req;
+    const { method, query: { id }, cookies } = req;
+
+    const token = cookies.token;
 
     dbConnect();
 
@@ -19,6 +21,9 @@ export default async function handler(req, res) {
 
     // CREATE PIZZA
     if (method === "PUT") {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not Authenticated!")
+        }
         try {
             const product = await Product.findByIdAndUpdate(id, req.body, {
                 new: true,
@@ -31,6 +36,9 @@ export default async function handler(req, res) {
 
     // DELETE PIZZA
     if (method === "DELETE") {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not Authenticated!")
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(200).json("The Product has been deleted!");
